add eventloop support to flower.
add an eventloop to the scheduler to handle IO and timer events. The
loop is based on libuv and is using the pyuv library.

While the other systems like eventlet and gevent pass all events in the
loop flower is running a loop task until all pending events are done. A
timer make sure to schedule the task and let the other tasks running using the
core scheduling system. Once all pending events are done the task loop
will be closed. If any other tasks need to handle new IO and timer event
a new task will be reopened. Only one loop exists at a time in a thread
and can be used by all tasks in this thread.