remove Timeout and fix timer.
The timeout features like it's desig for thread doesn't work very well
with our concurrency model. It will be replaced soon with channel
operations and the select feature taht you can find in go. See this
post speaking about it:

http://blog.golang.org/2010/09/go-concurrency-patterns-timing-out-and.html

Fix also the timer so we can process as fast as possible the timer. IT
is also now using only one task to handle them. With this change we
doesn't rely on the event loop for any timing operation.