refactor the code to handle udp that present a different interface
We need to fake the accept workflow in UDP since UDP doesn't really works
like TCP. On the other hand we still want to keep the advantages of this
protocol. So what we do is the following:

    1. on first accept we start to receive on the bound address
    2. on receive we check if this client is already handle by a
connection somewhere. If it's the case we send it the received data. If
not we create a new one that we link to the listener.

Which means that accept can happen in any task and receiving too.