fix race conditions.
We don't want to be stick in the uv task so we reschedule it as fast as
possible if this is the current task. While I'm here skip the runtime
error if main is the last task. At some point we may want to introduce
our own runtime wich would help a lot.